\name{setAsymptoticCovMat}
\alias{setAsymptoticCovMat}
\title{Set the asymptotic covariance matrix of a fitted HMM}
\description{This function sets the empirical asymptotic covariance matrix of the fitted HMM}
\usage{
setAsymptoticCovMat(HMMFit, asymptCovMat)
}
\arguments{
    \item{HMMFit}{a HMMFitClass object}
    \item{asymptCovMat}{The covariance matrix of the fitted model}
    
}
\value{The HMMFit object}



\seealso{asymptoticCovMat}
