\name{SunSpots}
\alias{SunSpots}

\docType{data}
\title{
SunSpots
}
\description{
Smoothed Monthly Sunspot.
}
\usage{data("SunSpots")}

\details{
Out of SunSpots.lrn' from 'D:/Subversion/lehre/Vorlesungen/TemporalesDataMining/01Text/05MarkovModelle'
}

\references{
...
}
\examples{
data(SunSpots)
str(SunSpots$SmothedMonSunspot)
str(SunSpots$TagNR)
}
\keyword{SunSpots}
