\name{weather}
\docType{data}
\alias{weather}
\title{Simulated discrete HMM}
\description{This data set contains 5 samples of discrete HMM}
\usage{data(weather)}
\format{R script}
\source{I took this example from Professor RD Boyle's tutorial, I used the same parameters and simulated the data with HMMSim.}
\references{\url{http://www.comp.leeds.ac.uk/roger/}}

\keyword{datasets}
